+++ app/helpers/users_helper.rb
@@ -7,8 +7,4 @@ module UsersHelper
   def about_tab(param_tab)
     return 'active' unless param_tab == "activity"
   end
-  
-  def deactivated_user_with_email(email)
-    User.only_deleted.where(email: email).first
-  end
 end
